refactor(key_distribution): Extract test payload in test_data module
